Add support for Syntastic
Syntastic [1] is a syntax checking plugin. It can be useful to introduce
additional linting info.

A new configuration `g:eleline_syntastic_raw` is introduced. When set to
`0` (default), the built-in format is used; when set to `1`, it calls
Syntastic's `SyntasticStatuslineFlag()` directly, which can be
configured by user via `g:syntastic_stl_format` [2].

[1] https://github.com/vim-syntastic/syntastic
[2] https://github.com/vim-syntastic/syntastic/blob/176f364f1b4acf894cdc248944f2d5506a37cc80/doc/syntastic.txt#L627-L643